+++ GuitarFX-Firmware/Audio/Effects/Inc/gain.h
@@ -5,7 +5,7 @@
 
 typedef audio_effect_t gain_effect_t;
 
-void gain_process(int32_t samples[restrict static 1], uint32_t frame_count);
+void gain_process(float samples[restrict static 1], uint32_t frame_count);
 void gain_reset(void);
 void gain_init(float gain);
